fix(api-members): adjustments to the GET /challenges/:challengeId/members route
