+++ website/server/controllers/api-v3/members.js
@@ -26,6 +26,7 @@ import {
   sanitizeText as sanitizeMessageText,
 } from '../../models/message';
 import highlightMentions from '../../libs/highlightMentions';
+import { handleGetMembersForChallenge } from '../../libs/challenges/handleGetMembersForChallenge';
 
 const { achievements } = common;
 
@@ -362,13 +363,7 @@ function _getMembersForItem (type) {
   };
 }
 
-async function _getMemberTasksFromChallenge (challengeId, memberId, user, res) {
-  const member = await User.findById(memberId).select(`${nameFields} challenges`).exec();
-  if (!member) throw new NotFound(res.t('userWithIDNotFound', { userId: memberId }));
-
-  const challenge = await Challenge.findById(challengeId).exec();
-  if (!challenge) throw new NotFound(res.t('challengeNotFound'));
-
+async function _getMemberTasksFromChallenge (challenge, member, user, res) {
   // optionalMembership is set to true because even if you're
   // not member of the group you may be able to access the challenge
   // for example if you've been booted from it, are the leader or a site admin
@@ -378,91 +373,13 @@ async function _getMemberTasksFromChallenge (challengeId, memberId, user, res) {
   if (!group || !challenge.canView(user, group)) throw new NotFound(res.t('challengeNotFound'));
   if (!challenge.isMember(member)) throw new NotFound(res.t('challengeMemberNotFound'));
 
-  const challengeTasks = await Tasks.Task.find({
-    userId: memberId,
-    'challenge.id': challengeId,
+  return Tasks.Task.find({
+    userId: member.id,
+    'challenge.id': challenge.id,
   })
-    .select('-tags') // We don't want to return the tags publicly TODO same for other data?
-    .exec();
-
-  return challengeTasks.map(task => {
-    task.checklist = []; // Clear checklists as they are private
-    return task.toJSON({ minimize: true });
-  });
-}
-
-async function _handleGetMembersForChallenge (req, res) {
-  req.checkParams('challengeId', res.t('challengeIdRequired')).notEmpty().isUUID();
-  req.checkQuery('lastId').optional().notEmpty().isUUID();
-  // Allow an arbitrary number of results (up to 60)
-  req.checkQuery('limit', res.t('groupIdRequired')).optional().notEmpty().isInt({ min: 1, max: 60 });
-
-  const validationErrors = req.validationErrors();
-  if (validationErrors) throw validationErrors;
-
-  const { challengeId } = req.params;
-  const { lastId } = req.query;
-  const { user } = res.locals;
-
-  const challenge = await Challenge.findById(challengeId).select('_id type leader group').exec();
-  if (!challenge) throw new NotFound(res.t('challengeNotFound'));
-
-  // optionalMembership is set to true because even
-  // if you're not member of the group you may be able to access the challenge
-  // for example if you've been booted from it, are the leader or a site admin
-  const group = await Group.getGroup({
-    user,
-    groupId: challenge.group,
-    fields: '_id type privacy',
-    optionalMembership: true,
-  });
-
-  if (!group || !challenge.canView(user, group)) throw new NotFound(res.t('challengeNotFound'));
-
-  const query = {};
-  let fields = nameFields;
-  // add computes stats to the member info when items and stats are available
-  let addComputedStats = false;
-  query.challenges = challenge._id;
-
-  if (req.query.includeAllPublicFields === 'true') {
-    fields = memberFields;
-    addComputedStats = true;
-  }
-
-  if (req.query.search) {
-    query['auth.local.username'] = { $regex: req.query.search };
-  }
-
-  if (lastId) query._id = { $gt: lastId };
-
-  let limit = req.query.limit ? Number(req.query.limit) : 30;
-
-  // Allow for all challenges members to be returned
-  if (req.query.includeAllMembers === 'true') {
-    limit = 0; // no limit
-  }
-
-  const members = await User
-    .find(query)
-    .sort({ _id: 1 })
-    .limit(limit)
-    .select(fields)
+    .select('-tags -checklist') // We don't want to return tags and checklists publicly
     .lean()
     .exec();
-
-  // manually call toJSON with minimize: true so empty paths aren't returned
-  members.forEach(member => User.transformJSONUser(member, addComputedStats));
-  if (req.query.includeTasks === 'true') {
-    const tasksPremises = members.map(
-      member => _getMemberTasksFromChallenge(challengeId, member.id, user, res),
-    );
-    const everyMemberTasks = await Promise.all(tasksPremises);
-    members.forEach((member, i) => {
-      member.tasks = everyMemberTasks[i];
-    });
-  }
-  res.respond(200, members);
 }
 
 /**
@@ -593,7 +510,7 @@ api.getMembersForChallenge = {
   method: 'GET',
   url: '/challenges/:challengeId/members',
   middlewares: [authWithHeaders()],
-  handler: _handleGetMembersForChallenge,
+  handler: handleGetMembersForChallenge,
 };
 
 /**
@@ -669,8 +586,10 @@ api.getChallengeMemberProgress = {
 
     const member = await User.findById(memberId).select(`${nameFields} challenges`).exec();
     if (!member) throw new NotFound(res.t('userWithIDNotFound', { userId: memberId }));
+    const challenge = await Challenge.findById(challengeId).exec();
+    if (!challenge) throw new NotFound(res.t('challengeNotFound'));
 
-    const challengeTasks = await _getMemberTasksFromChallenge(challengeId, memberId, user, res);
+    const challengeTasks = await _getMemberTasksFromChallenge(challenge, member, user, res);
 
     // manually call toJSON with minimize: true so empty paths aren't returned
     const response = member.toJSON({ minimize: true });
